# src/main/resources/static/api.yaml
swagger: "2.0"
info:
  description: "一个针对企业内部的组织架构、系统权限的管理平台。 业务原理可参考文章 [组织权限管理平台的设计与实现](https://mp.weixin.qq.com/s/h8aS_i3CTvkACsBYgqUipw)。可以使用Authorization `Authorization-token`来测试本应用。"
  version: "1.0.0"
  title: "组织权限管理"
  contact:
    email: "fanmeijian@aliyun.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8091/sparrow-platform"
basePath: "/"
...
schemes:
- "http"
paths:
  /oauth/password:
    post:
      tags:
      - "认证"
      summary: "修改密码"
      parameters:
      - name: "body"
        in: "body"
        schema:
          allOf: 
          - $ref: "#/definitions/Oauth"
          type: "object"
          properties:
            oldPassword:
              type: "string"
            secondPassword:
              type: "string"
      responses:
        "200":
          description: "成功"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Authorization: []
    put:
      tags:
      - "认证"
      summary: "重置密码"
      parameters:
      - name: "body"
        in: "body"
        schema:
          $ref: "#/definitions/Oauth"
      responses:
        "200":
          description: "成功"
          schema:
            $ref: "#/definitions/ApiResponse"
      security:
      - Authorization: []
...
securityDefinitions:
  Authorization:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  Oauth:
    type: "object"
    properties: 
      username:
        type: "string"
      password:
        type: "string"
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
      type:
        type: "string"
      message:
        type: "string"
...